# backend/docs/swagger.yaml
basePath: /api/
definitions:
...
  database.Elimination:
    properties:
      current_end:
        type: integer
      current_stage:
        type: integer
      group_id:
        type: integer
      id:
        type: integer
      medals:
        items:
          $ref: '#/definitions/database.Medal'
        type: array
      player_sets:
        items:
          $ref: '#/definitions/database.PlayerSet'
        type: array
      stages:
        items:
          $ref: '#/definitions/database.Stage'
        type: array
      team_size:
        type: integer
    type: object
...
  database.Match:
    properties:
      id:
        type: integer
      match_results:
        items:
          $ref: '#/definitions/database.MatchResult'
        type: array
      stage_id:
        type: integer
    type: object
...
  database.PlayerSet:
    properties:
      elimination_id:
        type: integer
      id:
        type: integer
      players:
        items:
          $ref: '#/definitions/database.Player'
        type: array
      rank:
        type: integer
      set_name:
        type: string
      total_score:
        type: integer
    type: object
...
  endpoint.PostElimination.PostEliminationData:
    properties:
      group_id:
        type: integer
      team_size:
        type: integer
    type: object
...
  endpoint.PostMatch.MatchData:
    properties:
      lane_numbers:
        items:
          type: integer
        type: array
      player_set_ids:
        items:
          type: integer
        type: array
      stage_id:
        type: integer
    type: object
...
  endpoint.PostStage.PostStageData:
    properties:
      elimination_id:
        type: integer
    type: object
...
  response.DeleteSuccessResponse:
    properties:
      message:
        example: 'Delete ID(1) : sth delete success'
        type: string
    type: object
  response.ErrorIdResponse:
    properties:
      error:
        example: 'invalid ID(1) : sth error'
        type: string
    type: object
  response.ErrorInternalErrorResponse:
    properties:
      error:
        example: 'sth need fix ID(1) : sth error'
        type: string
    type: object
...
  response.Nill:
    type: object
...
host: localhost:80
info:
  contact:
    name: NYCUArchery
    url: https://github.com/NYCUarchery
  description: Gin swagger
  license:
    name: no license yet
  title: Gin swagger
  version: "1.0"
paths:
...
  /elimination:
    post:
      consumes:
      - application/json
      description: Post one new Elimination data, and three medals
      parameters:
      - description: Elimination
        in: body
        name: Elimination
        required: true
        schema:
          $ref: '#/definitions/endpoint.PostElimination.PostEliminationData'
      produces:
      - application/json
      responses:
        "200":
          description: success, return one Elimination with new id
          schema:
            allOf:
            - $ref: '#/definitions/database.Elimination'
            - properties:
                medals:
                  $ref: '#/definitions/response.Nill'
                player_sets:
                  $ref: '#/definitions/response.Nill'
                stages:
                  $ref: '#/definitions/response.Nill'
              type: object
        "400":
          description: invalid group ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Create Elimination
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Create one Elimination
      tags:
      - Elimination
  /elimination/{id}:
    delete:
      consumes:
      - application/json
      description: Delete one Elimination by id, and related stages and matches
      parameters:
      - description: Elimination ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return delete success message
          schema:
            $ref: '#/definitions/response.DeleteSuccessResponse'
        "400":
          description: invalid Elimination ID, maybe not exist, or already deleted
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Delete Elimination
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Delete one Elimination, and related stages and matches
      tags:
      - Elimination
    get:
      description: Get only one Elimination by id
      parameters:
      - description: Elimination ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return one Elimination without related data
          schema:
            allOf:
            - $ref: '#/definitions/database.Elimination'
            - properties:
                medals:
                  $ref: '#/definitions/response.Nill'
                player_sets:
                  $ref: '#/definitions/response.Nill'
                stages:
                  $ref: '#/definitions/response.Nill'
              type: object
        "400":
          description: invalid Elimination ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Get Elimination
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show only one Elimination
      tags:
      - Elimination
  /elimination/currentend/minus/{id}:
    patch:
      description: Update one Elimination current end minus one by id
      parameters:
      - description: Elimination ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: success, return one Elimination with new current end
          schema:
            $ref: '#/definitions/response.Nill'
        "400":
          description: invalid Elimination ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Update Elimination CurrentStage
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Update one Elimination current end minus one
      tags:
      - Elimination
  /elimination/currentend/plus/{id}:
    patch:
      description: Update one Elimination current end plus one by id
      parameters:
      - description: Elimination ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: success, return one Elimination with new current end
          schema:
            $ref: '#/definitions/response.Nill'
        "400":
          description: invalid Elimination ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Update Elimination CurrentStage
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Update one Elimination current end plus one
      tags:
      - Elimination
  /elimination/currentstage/minus/{id}:
    patch:
      description: Update one Elimination current stage minus one by id
      parameters:
      - description: Elimination ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: success, return one Elimination with new current stage
          schema:
            $ref: '#/definitions/response.Nill'
        "400":
          description: invalid Elimination ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Update Elimination CurrentStage
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Update one Elimination current stage minus one
      tags:
      - Elimination
  /elimination/currentstage/plus/{id}:
    patch:
      description: Update one Elimination current stage plus one by id
      parameters:
      - description: Elimination ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: success, return one Elimination with new current stage
          schema:
            $ref: '#/definitions/response.Nill'
        "400":
          description: invalid Elimination ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Update Elimination CurrentStage
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Update one Elimination current stage plus one
      tags:
      - Elimination
  /elimination/match:
    post:
      consumes:
      - application/json
      description: |-
        Post one new Match data with 2 matchResults
        Each matchResults with 4 or 5 matchEnds with different teamSize
        Each matchEnds with 3, 4, 6 matchScores with different teamSize
        input PlayerSetIds should have 2 playerSets in the same elimination
        input LaneNumbers should have 2 laneNumbers for each playerset
      parameters:
      - description: Match
        in: body
        name: Match
        required: true
        schema:
          $ref: '#/definitions/endpoint.PostMatch.MatchData'
      produces:
      - application/json
      responses:
        "200":
          description: success, return one Match with new id
          schema:
            $ref: '#/definitions/database.Match'
        "400":
          description: invalid stage ID, maybe not exist, or player set id should
            be 2, lane numbers should be 2
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Create Match, MatchResult, MatchEnd,
            MatchScore, or get Stage
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Create one Match
      tags:
      - Elimination
  /elimination/match/scores/{matchid}:
    get:
      description: Get one Match with matchResults, matchEnds, scores, playerSets,
        players by id
      parameters:
      - description: Match ID
        in: path
        name: matchid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return one Match with all related data
          schema:
            $ref: '#/definitions/database.Match'
        "400":
          description: invalid Match ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Get Match
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show one Match with all related data
      tags:
      - Elimination
  /elimination/playersets/{id}:
    get:
      description: Get one Elimination with player sets by id
      parameters:
      - description: Elimination ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return one Elimination with player sets
          schema:
            allOf:
            - $ref: '#/definitions/database.Elimination'
            - properties:
                medals:
                  $ref: '#/definitions/response.Nill'
                player_sets:
                  allOf:
                  - $ref: '#/definitions/database.PlayerSet'
                  - properties:
                      players:
                        $ref: '#/definitions/response.Nill'
                    type: object
                stages:
                  $ref: '#/definitions/response.Nill'
              type: object
        "400":
          description: invalid Elimination ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Get Elimination
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show one Elimination with player sets
      tags:
      - Elimination
  /elimination/scores/{id}:
    get:
      description: Get one Elimination with stages, matches, matchResults, matchEnds,
        scores by id
      parameters:
      - description: Elimination ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return one Elimination with all scores
          schema:
            allOf:
            - $ref: '#/definitions/database.Elimination'
            - properties:
                medals:
                  $ref: '#/definitions/response.Nill'
                player_sets:
                  $ref: '#/definitions/response.Nill'
              type: object
        "400":
          description: invalid Elimination ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Get Elimination
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show one Elimination with all scores
      tags:
      - Elimination
  /elimination/stage:
    post:
      consumes:
      - application/json
      description: Post one new Stage data with new id
      parameters:
      - description: Stage
        in: body
        name: Stage
        required: true
        schema:
          $ref: '#/definitions/endpoint.PostStage.PostStageData'
      produces:
      - application/json
      responses:
        "200":
          description: success, return one Stage with new id
          schema:
            allOf:
            - $ref: '#/definitions/database.Stage'
            - properties:
                matchs:
                  $ref: '#/definitions/response.Nill'
              type: object
        "400":
          description: invalid elimination ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Create Stage
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Create one Stage
      tags:
      - Elimination
  /elimination/stages/matches/{id}:
    get:
      description: Get one Elimination with stages, matches by id
      parameters:
      - description: Elimination ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return one Elimination with stages, matches
          schema:
            allOf:
            - $ref: '#/definitions/database.Elimination'
            - properties:
                medals:
                  $ref: '#/definitions/response.Nill'
                player_sets:
                  $ref: '#/definitions/response.Nill'
                stages:
                  allOf:
                  - $ref: '#/definitions/database.Stage'
                  - properties:
                      matches:
                        allOf:
                        - $ref: '#/definitions/database.Match'
                        - properties:
                            MatchResult:
                              $ref: '#/definitions/response.Nill'
                          type: object
                    type: object
              type: object
        "400":
          description: invalid Elimination ID, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for Get Elimination
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Show one Elimination with stages, matches
      tags:
      - Elimination
...
swagger: "2.0"
...